+++ modulo2/apis-labefy/labefy/src/App.js
@@ -4,6 +4,7 @@ import Header from './Components/Header';
 import Footer from './Components/Footer';
 import Playlists from './Components/PlayLists';
 import PlayTracks from './Components/PlayTracks';
+import Menu from './Components/Menu'
 import styled from 'styled-components';
 import axios from 'axios';
 
@@ -14,6 +15,7 @@ border: 1px solid;
 `
 const Body = styled.div`
  border:2px solid navy;
+ display: flex;
 `
 
 
@@ -21,13 +23,17 @@ export default class App extends Component {
   render() {
     return (
       <Container>
-        <Header/>
+        <Header />
         <Body>
-        <Playlists/>
-        <PlayTracks/>
+          
+            <Menu />
+            
+            <PlayTracks />
+            
+          
         </Body>
-        <Footer/>
-        
+        <Footer />
+
       </Container>
     );
   }
