+++ modulo2/apis-labefy/labefy/src/Components/Footer.js
@@ -3,16 +3,18 @@ import { Component } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 
-
+const ContainerFooter = styled.div `
+background-color: purple;
+`
 
 
 export default class Footer extends Component {
   render() {
     return (
-      <div >
+      <ContainerFooter>
         <h1>FOOTER</h1>
         
-      </div>
+      </ContainerFooter>
     );
   }
 }
